Fixed Docker Bug
Fixed bug when using CI/CD the folder your is not /data. Updated code so it will
work in any folder in the docker image.

Updated tests to compare files instead of using pypandoc.
Mock the value of uuid, so we can predict what the image
file names will be.