:bug: Requiring `package.json` in script
Made some changes to the script, which
means it doesn't have to have `package.json` in the same folder.

This script was only tested for use within CI.
But it should work as a standalone script.
It now installs the node modules within the
home dir of the user and then references
the `mmdc` command from there.

Also made other changes such as adding a
version tag and updated documentation.